chequeamos rutas antes de borrar

+++ pineboolib/application/projectmodule.py
@@ -291,10 +291,12 @@ class Project(object):
 
             for root, dirs, files in os.walk(path._dir("cache/%s" % db_name), topdown=False):
                 for name in files:
-                    os.remove(os.path.join(root, name))
+                    if os.path.exists(os.path.join(root, name)):
+                        os.remove(os.path.join(root, name))
                 for name in dirs:
                     if name != "sqlite_database":
-                        os.rmdir(os.path.join(root, name))
+                        if os.path.exists(os.path.join(root, name)):
+                            os.rmdir(os.path.join(root, name))
 
         else:
             keep_images = settings.CONFIG.value("ebcomportamiento/keep_general_cache", False)
